+++ lib/abs_path_set.ml
@@ -1,0 +1,1 @@
+include Set.Make (Abs_path)