+++ CHANGELOG.md
@@ -6,7 +6,35 @@
 
 - Renamed `File_utils` to `Utils`
 
-- Added `Path_trie.remove` and `Path_trie_set.remove`
+- Added following functions to `Path_trie`:
+
+  - `is_empty`
+
+  - `remove`
+
+  - `merge`
+
+  - `union`
+
+  - `equal`
+
+  - `of_seq`
+
+- Added following functions to `Path_trie_set`:
+
+  - `is_empty`
+
+  - `remove`
+
+  - `union`
+
+  - `inter`
+
+  - `equal`
+
+  - `to_seq`
+
+  - `of_seq`
 
 ## 0.1.0